more prints, changed some fields from float to int

+++ schema.sql
@@ -9,11 +9,11 @@ CREATE TABLE appointments (
 CREATE TABLE test_results (
     name VARCHAR(255) NOT NULL,
     age INTEGER NOT NULL,
-    hemoglobin FLOAT NOT NULL,
-    red_blood_cell FLOAT NOT NULL,
-    white_blood_cell FLOAT NOT NULL,
-    platelets FLOAT NOT NULL,
-    neutrophils FLOAT NOT NULL,
+    hemoglobin INTEGER NOT NULL,
+    red_blood_cell INTEGER NOT NULL,
+    white_blood_cell INTEGER NOT NULL,
+    platelets INTEGER NOT NULL,
+    neutrophils INTEGER NOT NULL,
     signature TEXT NOT NULL,
     PRIMARY KEY (name)
 );
